# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baraq AI Service",
    "description": "Standalone AI platform for Fahes, Khota, Rasheed, Kholasa and Sada.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ai/v1/jobs": {
      "post": {
        "tags": [
          "AI Jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_ai_v1_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_JobAccepted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "AI Jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_ai_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_list_JobView__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "AI Jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_ai_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_JobView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/v1/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "AI Jobs"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_ai_v1_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIEnvelope_JobView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIEnvelope_JobAccepted_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobAccepted"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "APIEnvelope[JobAccepted]"
      },
      "APIEnvelope_JobView_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobView"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "APIEnvelope[JobView]"
      },
...
      "APIEnvelope_list_JobView__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/JobView"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/APIError"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "APIEnvelope[list[JobView]]"
      },
      "APIError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "APIError"
      },
      "Character": {
        "type": "string",
        "enum": [
          "fahes",
          "khota",
          "rasheed",
          "kholasa",
          "sada"
        ],
        "title": "Character"
      },
      "Citation": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "excerpt": {
            "type": "string",
            "maxLength": 1200,
            "minLength": 1,
            "title": "Excerpt"
          },
          "relevance_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_id",
          "excerpt"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "character": {
            "$ref": "#/components/schemas/Character"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url"
          },
          "estimated_wait_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Seconds"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "status",
          "task_type",
          "character",
          "status_url"
        ],
        "title": "JobAccepted"
      },
      "JobCreateRequest": {
        "properties": {
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "backend_request_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend Request Id"
          },
          "model_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fast",
                  "balanced",
                  "high_quality"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Tier"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_type",
          "payload",
          "idempotency_key"
        ],
        "title": "JobCreateRequest"
      },
      "JobOutputView": {
        "properties": {
          "output_id": {
            "type": "string",
            "format": "uuid",
            "title": "Output Id"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "validation_status": {
            "type": "string",
            "title": "Validation Status"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "groundedness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groundedness Score"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "provider_account": {
            "type": "string",
            "title": "Provider Account"
          },
          "materialized_resource_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialized Resource Type"
          },
          "materialized_resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materialized Resource Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "output_id",
          "result",
          "validation_status",
          "model_name",
          "provider_account"
        ],
        "title": "JobOutputView"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "retrieving",
          "generating",
          "validating",
          "materializing",
          "completed",
          "failed",
          "canceled"
        ],
        "title": "JobStatus"
      },
      "JobView": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "task_type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "character": {
            "$ref": "#/components/schemas/Character"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent"
          },
          "progress_message": {
            "type": "string",
            "title": "Progress Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOutputView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "task_type",
          "character",
          "status",
          "progress_percent",
          "progress_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobView"
      },
...
      "TaskType": {
        "type": "string",
        "enum": [
          "fahes_generate_quiz",
          "khota_generate_plan",
          "rasheed_recommend",
          "kholasa_summarize",
          "sada_transcribe",
          "source_ingest"
        ],
        "title": "TaskType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}